+++ CHANGELOG.md
@@ -6,7 +6,13 @@ separate changelogs for each crate were used. If you need to refer to these old 
 
 ## [Unreleased]
 
+### Changed
+
 - Bump Minimum Supported Rust Version (MSRV) to `1.60` ([#496](https://github.com/heroku/libcnb.rs/pull/496))
+
+### Added
+
+- Add new crate `libcnb-buildpack-commons` with common code that can be useful when implementing buildpacks with libcnb. ([#495](https://github.com/heroku/libcnb.rs/pull/495))
 
 ## [0.10.0] 2022-08-31